chore(136): split tests/test_mcp_tools.py into 5 files by feature
Split the 2,041-line test_mcp_tools.py into:
- test_mcp_validation.py (114 lines) — Pydantic input validation, schema aliases
- test_mcp_search_trace.py (326 lines) — search, find_references, find_column_usage, index_status, trace_dependencies, trace_column_lineage
- test_mcp_pr_impact.py (486 lines) — pr_impact end-to-end + reindex thread-safety
- test_mcp_reindex_files.py (379 lines) — reindex_files + debounce patterns
- test_mcp_schema_context.py (761 lines) — get_schema, check_impact, get_context

All 74 tests still pass. No behavior changes.
